# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: QR Menu System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products
      parameters:
        - name: categoryId
          in: query
          schema:
            type: ["integer", "null"]
        - name: activeOnly
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of products with full data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductFull"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductFull"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product with full data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductFull"
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductFull"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /products/reorder:
    post:
      operationId: reorderProducts
      tags: [products]
      summary: Reorder products
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReorderInput"
      responses:
        "200":
          description: Reordered
  /products/{id}/view:
    post:
      operationId: trackProductView
      tags: [products]
      summary: Record a product view
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ViewEventInput"
      responses:
        "200":
          description: Recorded

  # Public menu
...
  /storage/uploads/request-url:
    post:
      tags: [Storage]
      operationId: requestUploadUrl
      summary: Request a presigned URL for file upload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadUrlRequest"
      responses:
        "200":
          description: Presigned upload URL generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadUrlResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Failed to generate URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /storage/uploads/confirm:
    post:
      tags: [Storage]
      operationId: confirmUpload
      summary: Confirm an upload and set object to public ACL
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfirmUploadInput"
      responses:
        "200":
          description: Upload confirmed; returns public serving URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConfirmUploadResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
  /storage/objects/{objectPath}:
    get:
      tags: [Storage]
      operationId: getStorageObject
      summary: Serve an object entity
      parameters:
        - name: objectPath
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Object content
          content:
            "*/*":
              schema:
                type: string
                format: binary
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
components:
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    ProductTranslation:
      type: object
      required: [id, productId, languageCode, name]
      properties:
        id:
          type: integer
        productId:
          type: integer
        languageCode:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        ingredients:
          type: ["string", "null"]
        allergenNote:
          type: ["string", "null"]
        specialNote:
          type: ["string", "null"]
    ProductTranslationInput:
      type: object
      required: [languageCode, name]
      properties:
        languageCode:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        ingredients:
          type: ["string", "null"]
        allergenNote:
          type: ["string", "null"]
        specialNote:
          type: ["string", "null"]
    ProductNutrition:
      type: object
      properties:
        calories:
          type: ["number", "null"]
        energy:
          type: ["number", "null"]
        protein:
          type: ["number", "null"]
        carbs:
          type: ["number", "null"]
        fat:
          type: ["number", "null"]
        isAiGenerated:
          type: boolean
    ProductFull:
      type: object
      required: [id, categoryId, price, isActive, sortOrder, translations, allergens, viewCount]
      properties:
        id:
          type: integer
        categoryId:
          type: ["integer", "null"]
        price:
          type: ["number", "null"]
        imagePath:
          type: ["string", "null"]
        isActive:
          type: boolean
        isFeatured:
          type: boolean
        sortOrder:
          type: integer
        viewCount:
          type: integer
        translations:
          type: array
          items:
            $ref: "#/components/schemas/ProductTranslation"
        nutrition:
          $ref: "#/components/schemas/ProductNutrition"
        allergens:
          type: array
          items:
            type: string
    ProductInput:
      type: object
      required: [translations]
      properties:
        categoryId:
          type: ["integer", "null"]
        price:
          type: ["number", "null"]
        imagePath:
          type: ["string", "null"]
        isActive:
          type: boolean
        isFeatured:
          type: boolean
        sortOrder:
          type: integer
        translations:
          type: array
          items:
            $ref: "#/components/schemas/ProductTranslationInput"
        nutrition:
          $ref: "#/components/schemas/ProductNutrition"
        allergens:
          type: array
          items:
            type: string
    ProductUpdate:
      type: object
      properties:
        categoryId:
          type: ["integer", "null"]
        price:
          type: ["number", "null"]
        imagePath:
          type: ["string", "null"]
        isActive:
          type: boolean
        isFeatured:
          type: boolean
        sortOrder:
          type: integer
        translations:
          type: array
          items:
            $ref: "#/components/schemas/ProductTranslationInput"
        nutrition:
          $ref: "#/components/schemas/ProductNutrition"
        allergens:
          type: array
          items:
            type: string
    ReorderInput:
      type: object
      required: [ids]
      properties:
        ids:
          type: array
          items:
            type: integer
    ViewEventInput:
      type: object
      properties:
        languageCode:
          type: ["string", "null"]
...
    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name:
          type: string
          minLength: 1
        size:
          type: integer
          minimum: 1
        contentType:
          type: string
          minLength: 1
    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL:
          type: string
          format: uri
        objectPath:
          type: string
        metadata:
          $ref: "#/components/schemas/UploadUrlRequest"
    ConfirmUploadInput:
      type: object
      required: [objectPath]
      properties:
        objectPath:
          type: string
          description: The object path returned from the presigned URL request
    ConfirmUploadResponse:
      type: object
      required: [servingUrl, objectPath]
      properties:
        servingUrl:
          type: string
          description: Public URL for serving the uploaded file
        objectPath:
          type: string
...